Added activation collection utility
Added Connectivity.py file to assign all code related to collecting connectivity values. This commit implements the activation calculation functions needed to get the activation values of the expert adapters for a given subnetwork on a provided data loader. This provides the raw activation values needed for calculating connectivity values.